requirements.txt
openai>=1.37.0
faiss-cpu==1.8.0.post1
sentence-transformers==2.6.1
pypdf>=4.2.0
pandas>=2.2.2
pyarrow>=16.1.0
numpy>=1.26.4
python-dotenv>=1.0.1
tqdm>=4.66.4
rich>=13.7.1
click>=8.1.7
pydantic>=2.8.0
pdfminer.six>=20231228